+++ bio_utils/tests/test_subject_sequence_retriever.py
@@ -1,0 +1,73 @@
+#! /usr/bin/env python
+
+"""Test bio_utils' subject_sequence_retriever
+
+Copyright:
+
+    test_subject_sequence_retriever.py test bio_utils' subject_sequence_retriever
+    Copyright (C) 2015  William Brazelton, Alex Hyer
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+
+from ..blast_tools import subject_sequence_retriever
+import os
+
+__author__ = 'Alex Hyer'
+__email__ = 'theonehyer@gmail.com'
+__license__ = 'GPLv3'
+__maintainer__ = 'Alex Hyer'
+__status__ = 'Production'
+__version__ = '1.0.0'
+
+
+def test_subject_sequence_retiever():
+    """Test bio_utils' subject_sequence_retriever"""
+
+    # Store properly formatted B6/M8 data
+    b6_data = 'query1\tsubject1\t86.03\t10\t3\t1\t15\t5\t' \
+              '15\t5\t1E-5\t1890{0}' \
+              'query2\tsubject2\t95.46\t23\t5\t7\t10\t33\t' \
+              '10\t33\t3E-0\t1219{0}' \
+              'query3\tsubject3\t85.46\t13\t2\t5\t10\t23\t' \
+              '10\t23\t1E-37\t1219'.format(os.linesep)
+
+    b6_handle = iter(b6_data.split(os.linesep))
+
+    # Store properly formatted FASTA data
+    fasta_data = '>subject1{0}AGGCTAGGCTAGCTGGTCAAGGCT{0}' \
+                 '>subject3 description{0}AAGGGCGCTTGCGAGCTTAGCTAGAGCTAGGCTA' \
+                 .format(os.linesep)
+
+    fasta_handle = iter(fasta_data.split(os.linesep))
+
+    entries = []
+    for entry in subject_sequence_retriever(fasta_handle, b6_handle, 1):
+        entries.append(entry)
+
+    assert len(entries) == 2  # Ensure high E-value entry dropped
+
+    # Test first entry with reverse sequence alignment (start: 15, end: 5)
+    assert entries[0].id == 'subject1'
+    assert entries[0].description == 'E-value: 1E-5'
+    assert entries[0].sequence == 'TCGATCGGAT'
+    assert entries[0].write() == '>subject1 E-value: 1E-5{0}' \
+                                 'TCGATCGGAT{0}'.format(os.linesep)
+
+    # Test second entry with description
+    assert entries[1].id == 'subject3'
+    assert entries[1].description == 'description E-value: 1E-37'
+    assert entries[1].sequence == 'TGCGAGCTTAGCT'
+    assert entries[1].write() == '>subject3 description E-value: 1E-37{0}' \
+                                 'TGCGAGCTTAGCT{0}'.format(os.linesep)